+++ ios/Classes/signalr_flutter.h
@@ -1,0 +1,8 @@
+//
+//  signalr_flutter.h
+//  signalr_flutter
+//
+//  Created by Ayon Das on 07/11/21.
+//
+
+#import "SignalrApi.h"