+++ test/signalr_flutter_test.dart
@@ -1,27 +1,14 @@
 import 'package:flutter/services.dart';
 import 'package:flutter_test/flutter_test.dart';
-import 'package:signalr_flutter/signalr_flutter.dart';
 
-void main() async {
-  const MethodChannel channel = MethodChannel('signalR');
-  final SignalR signalR = SignalR('Url', "hubName");
+void main() {
+  const MethodChannel channel = MethodChannel('signalr_flutter');
 
   TestWidgetsFlutterBinding.ensureInitialized();
 
   setUp(() {
     channel.setMockMethodCallHandler((MethodCall methodCall) async {
-      switch (methodCall.method) {
-        case "connectToServer":
-          return true;
-        case "invokeServerMethod":
-          return <String, dynamic>{
-            'baseUrl': "123",
-            "hubName": "456",
-          };
-        default:
-          return PlatformException(
-              code: "Error", message: "No implementation found for method ${methodCall.method}");
-      }
+      return '42';
     });
   });
 
@@ -29,18 +16,5 @@ void main() async {
     channel.setMockMethodCallHandler(null);
   });
 
-  group('SignalR Test', () {
-    test('Connect SignalR', () async {
-      final result = await signalR.connect();
-      expect(result, true);
-    });
-
-    test('Invoke Server Method', () async {
-      final result = await signalR.invokeMethod("methodName", arguments: null);
-      expect(result, <String, dynamic>{
-        'baseUrl': "123",
-        "hubName": "456",
-      });
-    });
-  });
+  test('getPlatformVersion', () async {});
 }